feature: start time & end time.

+++ api/v1/form/form_service_plan.go
@@ -8,7 +8,7 @@ type ServicePlanForm struct {
 	PlanEnabled     bool   // 启用中
 	PlanPrice       float32
 
-	ServingDays int16
+	ServingDays int
 
 	BundledTrafficPlanID  uint
 	AvailableTrafficPlans []uint
